BIGC-618 moved tools directory to root of lib
