+++ extension/lib/tools/TimeLogger.js
